# MM-39918: mmctl: list archived private channels in channel list command with system admin user