--- a/server/channels/store/store.go
+++ b/server/channels/store/store.go
@@ -202,7 +202,7 @@ type ChannelStore interface {
 	GetByNameIncludeDeleted(teamID string, name string, allowFromCache bool) (*model.Channel, error)
 	GetByNamesIncludeDeleted(teamID string, names []string, allowFromCache bool) ([]*model.Channel, error)
 	GetDeletedByName(teamID string, name string) (*model.Channel, error)
-	GetDeleted(teamID string, offset int, limit int, userID string) (model.ChannelList, error)
+	GetDeleted(teamID string, offset int, limit int, userID string, includePrivate bool) (model.ChannelList, error)
 	GetChannels(teamID, userID string, opts *model.ChannelSearchOpts) (model.ChannelList, error)
 	GetChannelsByUser(userID string, includeDeleted bool, lastDeleteAt, pageSize int, fromChannelID string) (model.ChannelList, error)
 	GetAllChannelMemberIdsByChannelId(id string) ([]string, error)