--- a/server/channels/store/storetest/channel_store.go
+++ b/server/channels/store/storetest/channel_store.go
@@ -1006,7 +1006,7 @@ func testChannelStoreGetDeleted(t *testing.T, rctx request.CTX, ss store.Store) 
 	err := ss.Channel().Delete(o1.Id, model.GetMillis())
 	require.NoError(t, err, "channel should have been deleted")
 
-	list, nErr := ss.Channel().GetDeleted(o1.TeamId, 0, 100, userID)
+	list, nErr := ss.Channel().GetDeleted(o1.TeamId, 0, 100, userID, false)
 	require.NoError(t, nErr, nErr)
 	require.Len(t, list, 1, "wrong list")
 	require.Equal(t, o1.Name, list[0].Name, "missing channel")
@@ -1019,7 +1019,7 @@ func testChannelStoreGetDeleted(t *testing.T, rctx request.CTX, ss store.Store) 
 	_, nErr = ss.Channel().Save(rctx, &o2, -1)
 	require.NoError(t, nErr)
 
-	list, nErr = ss.Channel().GetDeleted(o1.TeamId, 0, 100, userID)
+	list, nErr = ss.Channel().GetDeleted(o1.TeamId, 0, 100, userID, false)
 	require.NoError(t, nErr, nErr)
 	require.Len(t, list, 1, "wrong list")
 
@@ -1035,15 +1035,15 @@ func testChannelStoreGetDeleted(t *testing.T, rctx request.CTX, ss store.Store) 
 	err = ss.Channel().Delete(o3.Id, model.GetMillis())
 	require.NoError(t, err, "channel should have been deleted")
 
-	list, nErr = ss.Channel().GetDeleted(o1.TeamId, 0, 100, userID)
+	list, nErr = ss.Channel().GetDeleted(o1.TeamId, 0, 100, userID, false)
 	require.NoError(t, nErr, nErr)
 	require.Len(t, list, 2, "wrong list length")
 
-	list, nErr = ss.Channel().GetDeleted(o1.TeamId, 0, 1, userID)
+	list, nErr = ss.Channel().GetDeleted(o1.TeamId, 0, 1, userID, false)
 	require.NoError(t, nErr, nErr)
 	require.Len(t, list, 1, "wrong list length")
 
-	list, nErr = ss.Channel().GetDeleted(o1.TeamId, 1, 1, userID)
+	list, nErr = ss.Channel().GetDeleted(o1.TeamId, 1, 1, userID, false)
 	require.NoError(t, nErr, nErr)
 	require.Len(t, list, 1, "wrong list length")
 }