--- a/server/channels/store/storetest/mocks/ChannelStore.go
+++ b/server/channels/store/storetest/mocks/ChannelStore.go
@@ -1210,9 +1210,9 @@ func (_m *ChannelStore) GetChannelsWithUnreadsAndWithMentions(ctx context.Contex
 	return r0, r1, r2, r3
 }
 
-// GetDeleted provides a mock function with given fields: teamID, offset, limit, userID
-func (_m *ChannelStore) GetDeleted(teamID string, offset int, limit int, userID string) (model.ChannelList, error) {
-	ret := _m.Called(teamID, offset, limit, userID)
+// GetDeleted provides a mock function with given fields: teamID, offset, limit, userID, includePrivate
+func (_m *ChannelStore) GetDeleted(teamID string, offset int, limit int, userID string, includePrivate bool) (model.ChannelList, error) {
+	ret := _m.Called(teamID, offset, limit, userID, includePrivate)
 
 	if len(ret) == 0 {
 		panic("no return value specified for GetDeleted")
@@ -1220,19 +1220,19 @@ func (_m *ChannelStore) GetDeleted(teamID string, offset int, limit int, userID 
 
 	var r0 model.ChannelList
 	var r1 error
-	if rf, ok := ret.Get(0).(func(string, int, int, string) (model.ChannelList, error)); ok {
-		return rf(teamID, offset, limit, userID)
+	if rf, ok := ret.Get(0).(func(string, int, int, string, bool) (model.ChannelList, error)); ok {
+		return rf(teamID, offset, limit, userID, includePrivate)
 	}
-	if rf, ok := ret.Get(0).(func(string, int, int, string) model.ChannelList); ok {
-		r0 = rf(teamID, offset, limit, userID)
+	if rf, ok := ret.Get(0).(func(string, int, int, string, bool) model.ChannelList); ok {
+		r0 = rf(teamID, offset, limit, userID, includePrivate)
 	} else {
 		if ret.Get(0) != nil {
 			r0 = ret.Get(0).(model.ChannelList)
 		}
 	}
 
-	if rf, ok := ret.Get(1).(func(string, int, int, string) error); ok {
-		r1 = rf(teamID, offset, limit, userID)
+	if rf, ok := ret.Get(1).(func(string, int, int, string, bool) error); ok {
+		r1 = rf(teamID, offset, limit, userID, includePrivate)
 	} else {
 		r1 = ret.Error(1)
 	}