#### Summary
Add parameter in GetDeletedChannels to include private channels if required and enable it for users with PermissionManageSystem

I'm trying to add tests but I need a bit of time to understand how it works

#### Ticket Link
Fixes https://github.com/mattermost/mattermost/issues/19913
Jira https://mattermost.atlassian.net/browse/MM-39918

#### Release Note
```release-note
NONE
```