@KuSh 2024-10-14
> @KuSh The tests are failing. Can you please take a look?

I can't reproduce locally :

```
❯ ./bin/gotestsum --packages="github.com/mattermost/mattermost/server/v8/channels/store/sqlstore"
✓  channels/store/sqlstore (7m36.817s)

=== Skipped
=== SKIP: channels/store/sqlstore TestFileInfoStore/MySQL/GetUptoNSizeFileTime (0.00s)
    file_info_store.go:873: MM-53905

=== SKIP: channels/store/sqlstore TestFileInfoStore/PostgreSQL/GetUptoNSizeFileTime (0.00s)
    file_info_store.go:873: MM-53905

=== SKIP: channels/store/sqlstore TestThreadStore/MySQL/ThreadStorePopulation/Get_unread_reply_counts_for_thread (0.00s)
    thread_store.go:421: MM-41797

=== SKIP: channels/store/sqlstore TestThreadStore/PostgreSQL/ThreadStorePopulation/Get_unread_reply_counts_for_thread (0.00s)
    thread_store.go:421: MM-41797

=== SKIP: channels/store/sqlstore TestUserStore/MySQL/GetProfilesNotInTeam/etag_for_profiles_not_in_team_1_after_u3_added_to_team_2 (0.00s)
    user_store.go:4790: 

=== SKIP: channels/store/sqlstore TestUserStore/PostgreSQL/GetProfilesNotInTeam/etag_for_profiles_not_in_team_1_after_u3_added_to_team_2 (0.00s)
    user_store.go:4790: 

DONE 5481 tests, 6 skipped in 493.378s

❯ ./bin/gotestsum --packages="github.com/mattermost/mattermost/server/v8/channels/store/storetest" -- -race $GOFLAGS -timeout=90m
✓  channels/store/storetest (2.087s)

DONE 1 tests in 14.031s
```

Does CI use a particular configuration ? Is it reproducible locally ?