@hanzei 2024-10-14
Hm, I can't reproduce the failure either. Let me merge `master` into your PR and see if that fixes the CI issue.