@KuSh 2024-10-14
> Hm, I can't reproduce the failure either. Let me merge `master` into your PR and see if that fixes the CI issue.

It doesn't seem to have helped :/
I'll try to follow exactly the steps the CI takes to see if I can reproduce locally. Meanwhile If you have any leads, I'm all ears!