@hanzei 2024-10-15
@KuSh You can run the tests locally using `go test -timeout 300s -run ^TestChannelStore$ ./channels/store/sqlstore -v`. With the command, they also error on my local setup.