@esarafianou 2024-10-25
@agarciamontoro since we changed the SQL query, can you re-review please?