@agarciamontoro 2024-10-25
Ooooh, that's a great improvement indeed, thanks! Let me double-check that the query plan is similar in both, will approve later if no performance issues comes up :)