@agarciamontoro 2024-10-25
We've ran an `EXPLAIN ANALYZE` on both the old and new queries on Community, using my user ID, and things seem to even be slightly better for the new query: the cost, the number of buffers hit, and the times are all lower in the new query :tada:

This is a specific example with a specific dataset, so things may change in other scenarios, but these results give me some peace of mind, I'm no longer concerned about potential performance regressions for the change. Thanks, @KuSh and @esarafianou, this looks good to me!

#### Old query
```
+------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
|QUERY PLAN                                                                                                                                                                                                                                                                                                                                                                                                        |
+------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
|Limit  (cost=6732.67..6732.92 rows=100 width=1674) (actual time=18.186..18.202 rows=100 loops=1)                                                                                                                                                                                                                                                                                                                  |
|  Buffers: shared hit=6047                                                                                                                                                                                                                                                                                                                                                                                        |
|  ->  Sort  (cost=6732.67..6737.79 rows=2046 width=1674) (actual time=18.185..18.191 rows=100 loops=1)                                                                                                                                                                                                                                                                                                            |
|        Sort Key: channels.displayname                                                                                                                                                                                                                                                                                                                                                                            |
|        Sort Method: top-N heapsort  Memory: 100kB                                                                                                                                                                                                                                                                                                                                                                |
|        Buffers: shared hit=6047                                                                                                                                                                                                                                                                                                                                                                                  |
|        ->  HashAggregate  (cost=6634.02..6654.48 rows=2046 width=1674) (actual time=17.084..17.417 rows=955 loops=1)                                                                                                                                                                                                                                                                                             |
|              Group Key: channels.id, channels.createat, channels.updateat, channels.deleteat, channels.teamid, channels.type, channels.displayname, channels.name, channels.header, channels.purpose, channels.lastpostat, channels.totalmsgcount, channels.extraupdateat, channels.creatorid, channels.schemeid, channels.groupconstrained, channels.shared, channels.totalmsgcountroot, channels.lastrootpostat|
|              Batches: 1  Memory Usage: 612kB                                                                                                                                                                                                                                                                                                                                                                     |
|              Buffers: shared hit=6044                                                                                                                                                                                                                                                                                                                                                                            |
|              ->  Append  (cost=0.00..6536.83 rows=2046 width=1674) (actual time=0.010..16.130 rows=955 loops=1)                                                                                                                                                                                                                                                                                                  |
|                    Buffers: shared hit=6044                                                                                                                                                                                                                                                                                                                                                                      |
|                    ->  Seq Scan on channels  (cost=0.00..4561.94 rows=2045 width=186) (actual time=0.010..12.071 rows=906 loops=1)                                                                                                                                                                                                                                                                               |
|                          Filter: ((deleteat <> 0) AND (type <> 'P'::channel_type) AND (((teamid)::text = 'REDACTED'::text) OR ((teamid)::text = ''::text)))                                                                                                                                                                                                                                                      |
|                          Rows Removed by Filter: 83500                                                                                                                                                                                                                                                                                                                                                           |
|                          Buffers: shared hit=2876                                                                                                                                                                                                                                                                                                                                                                |
|                    ->  Nested Loop  (cost=0.84..1944.20 rows=1 width=186) (actual time=0.218..3.959 rows=49 loops=1)                                                                                                                                                                                                                                                                                             |
|                          Buffers: shared hit=3168                                                                                                                                                                                                                                                                                                                                                                |
|                          ->  Index Only Scan using idx_channelmembers_user_id_channel_id_last_viewed_at on channelmembers  (cost=0.42..240.74 rows=785 width=27) (actual time=0.041..0.448 rows=672 loops=1)                                                                                                                                                                                                     |
|                                Index Cond: (userid = 'REDACTED'::text)                                                                                                                                                                                                                                                                                                                                           |
|                                Heap Fetches: 474                                                                                                                                                                                                                                                                                                                                                                 |
|                                Buffers: shared hit=256                                                                                                                                                                                                                                                                                                                                                           |
|                          ->  Index Scan using channels_pkey on channels channels_1  (cost=0.42..2.17 rows=1 width=186) (actual time=0.005..0.005 rows=0 loops=672)                                                                                                                                                                                                                                               |
|                                Index Cond: ((id)::text = (channelmembers.channelid)::text)                                                                                                                                                                                                                                                                                                                       |
|                                Filter: ((deleteat <> 0) AND (type = 'P'::channel_type) AND (((teamid)::text = 'REDACTED'::text) OR ((teamid)::text = ''::text)))                                                                                                                                                                                                                                                 |
|                                Rows Removed by Filter: 1                                                                                                                                                                                                                                                                                                                                                         |
|                                Buffers: shared hit=2912                                                                                                                                                                                                                                                                                                                                                          |
|Planning:                                                                                                                                                                                                                                                                                                                                                                                                         |
|  Buffers: shared hit=553                                                                                                                                                                                                                                                                                                                                                                                         |
|Planning Time: 1.473 ms                                                                                                                                                                                                                                                                                                                                                                                           |
|Execution Time: 18.329 ms                                                                                                                                                                                                                                                                                                                                                                                         |
+------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
```
#### New query
```
+---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
|QUERY PLAN                                                                                                                                                                                               |
+---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
|Limit  (cost=5304.41..5304.66 rows=100 width=186) (actual time=12.343..12.359 rows=100 loops=1)                                                                                                          |
|  Buffers: shared hit=3132                                                                                                                                                                               |
|  ->  Sort  (cost=5304.41..5309.53 rows=2048 width=186) (actual time=12.342..12.349 rows=100 loops=1)                                                                                                    |
|        Sort Key: channels.displayname                                                                                                                                                                   |
|        Sort Method: top-N heapsort  Memory: 92kB                                                                                                                                                        |
|        Buffers: shared hit=3132                                                                                                                                                                         |
|        ->  Seq Scan on channels  (cost=242.71..5226.13 rows=2048 width=186) (actual time=0.017..11.682 rows=955 loops=1)                                                                                |
|              Filter: ((deleteat <> 0) AND (((teamid)::text = 'REDACTED'::text) OR ((teamid)::text = ''::text)) AND ((type <> 'P'::channel_type) OR ((type = 'P'::channel_type) AND (hashed SubPlan 1))))|
|              Rows Removed by Filter: 83451                                                                                                                                                              |
|              Buffers: shared hit=3132                                                                                                                                                                   |
|              SubPlan 1                                                                                                                                                                                  |
|                ->  Index Only Scan using idx_channelmembers_user_id_channel_id_last_viewed_at on channelmembers  (cost=0.42..240.74 rows=785 width=27) (actual time=0.032..0.429 rows=672 loops=1)      |
|                      Index Cond: (userid = REDACTED::text)                                                                                                                                              |
|                      Heap Fetches: 474                                                                                                                                                                  |
|                      Buffers: shared hit=256                                                                                                                                                            |
|Planning Time: 0.201 ms                                                                                                                                                                                  |
|Execution Time: 12.413 ms                                                                                                                                                                                |
+---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
```